fix mv简介item的高度
