+++ lib/pages/page_comment.dart
@@ -1,7 +1,6 @@
 import 'package:async/async.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:intl/intl.dart';
 import 'package:quiet/model/playlist_detail.dart';
 import 'package:quiet/part/part.dart';
 import 'package:quiet/repository/netease.dart';
@@ -657,29 +656,6 @@ enum CommentType {
 
   ///video comments
   video
-}
-
-///format milliseconds to local string
-String getFormattedTime(int milliseconds) {
-  var dateTime = DateTime.fromMillisecondsSinceEpoch(milliseconds);
-  var now = DateTime.now();
-
-  var diff = Duration(
-      milliseconds:
-          now.millisecondsSinceEpoch - dateTime.millisecondsSinceEpoch);
-  if (diff.inMinutes < 1) {
-    return "刚刚";
-  }
-  if (diff.inMinutes <= 60) {
-    return "${diff.inMinutes}分钟前";
-  }
-  if (diff.inHours <= 24) {
-    return "${diff.inHours}小时前";
-  }
-  if (diff.inDays <= 5) {
-    return "${diff.inDays}天前";
-  }
-  return DateFormat("y年M月d日").format(dateTime);
 }
 
 ///like or unlike a comment
