+++ lib/pages/page_mv_detail.dart
@@ -99,12 +99,12 @@ class _MvDetailPageState extends State<_MvDetailPage> {
                         delegate: SliverChildListDelegate([
                           _MvInformationSection(data: widget.result['data'])
                         ]),
-                        itemExtent: 76),
+                        itemExtent: _MvInformationSection.height),
                     SliverFixedExtentList(
                         delegate: SliverChildListDelegate([
                           _MvActionsSection(),
                         ]),
-                        itemExtent: 62),
+                        itemExtent: _MvActionsSection.height),
                   ];
                 },
                 body: Loader(
@@ -285,30 +285,35 @@ class _MvInformationSection extends StatelessWidget {
 
   const _MvInformationSection({Key key, this.data}) : super(key: key);
 
+  static const double height = 70;
+
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.all(8),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Text(data['name'],
-              style: Theme.of(context)
-                  .textTheme
-                  .title
-                  .copyWith(fontWeight: FontWeight.bold)),
-          DefaultTextStyle(
-            style: TextStyle(color: Colors.grey),
-            child: Row(
-              children: <Widget>[
-                Text('发布: ${data['publishTime']}'),
-                VerticalDivider(color: Theme.of(context).dividerColor),
-                Text('播放: ${getFormattedNumber(data['playCount'])}')
-              ],
-            ),
-          )
-        ],
+    return Container(
+      height: 70,
+      padding: const EdgeInsets.symmetric(horizontal: 8),
+      child: Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Text(data['name'],
+                style: Theme.of(context)
+                    .textTheme
+                    .title
+                    .copyWith(fontWeight: FontWeight.bold)),
+            DefaultTextStyle(
+              style: TextStyle(color: Colors.grey),
+              child: Row(
+                children: <Widget>[
+                  Text('发布: ${data['publishTime']}'),
+                  VerticalDivider(color: Theme.of(context).dividerColor),
+                  Text('播放: ${getFormattedNumber(data['playCount'])}')
+                ],
+              ),
+            )
+          ],
+        ),
       ),
     );
   }
@@ -316,62 +321,66 @@ class _MvInformationSection extends StatelessWidget {
 
 ///mv 点赞/收藏/评论/分享
 class _MvActionsSection extends StatelessWidget {
+  static final double height = 72;
+
   @override
   Widget build(BuildContext context) {
     final data = MvPlayerModel.of(context).mvData;
     return DividerWrapper(
-      extent: 10,
       child: ButtonTheme(
         textTheme: ButtonTextTheme.accent,
         colorScheme: ColorScheme.light(secondary: Colors.black54),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: <Widget>[
-            FlatButton(
-                onPressed: () {},
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    const SizedBox(height: 4.0),
-                    Icon(Icons.thumb_up),
-                    const SizedBox(height: 4.0),
-                    Text('${data['likeCount']}'),
-                  ],
-                )),
-            FlatButton(
-                onPressed: () {},
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    const SizedBox(height: 4.0),
-                    Icon(Icons.add_box),
-                    const SizedBox(height: 4.0),
-                    Text('${data['subCount']}'),
-                  ],
-                )),
-            FlatButton(
-                onPressed: () {},
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    const SizedBox(height: 4.0),
-                    Icon(Icons.comment),
-                    const SizedBox(height: 4.0),
-                    Text('${data['commentCount']}'),
-                  ],
-                )),
-            FlatButton(
-                onPressed: () {},
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    const SizedBox(height: 4.0),
-                    Icon(Icons.share),
-                    const SizedBox(height: 4.0),
-                    Text('${data['shareCount']}'),
-                  ],
-                )),
-          ],
+        child: Container(
+          height: height,
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: <Widget>[
+              FlatButton(
+                  onPressed: () {},
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: <Widget>[
+                      const SizedBox(height: 4.0),
+                      Icon(Icons.thumb_up),
+                      const SizedBox(height: 4.0),
+                      Text('${data['likeCount']}'),
+                    ],
+                  )),
+              FlatButton(
+                  onPressed: () {},
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: <Widget>[
+                      const SizedBox(height: 4.0),
+                      Icon(Icons.add_box),
+                      const SizedBox(height: 4.0),
+                      Text('${data['subCount']}'),
+                    ],
+                  )),
+              FlatButton(
+                  onPressed: () {},
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: <Widget>[
+                      const SizedBox(height: 4.0),
+                      Icon(Icons.comment),
+                      const SizedBox(height: 4.0),
+                      Text('${data['commentCount']}'),
+                    ],
+                  )),
+              FlatButton(
+                  onPressed: () {},
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: <Widget>[
+                      const SizedBox(height: 4.0),
+                      Icon(Icons.share),
+                      const SizedBox(height: 4.0),
+                      Text('${data['shareCount']}'),
+                    ],
+                  )),
+            ],
+          ),
         ),
       ),
     );
